apps/api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
websockets
psutil
